Fix .md table pipe parsing
The regex that populates "matches" gets confused when a markdown table pipe ("|") is directly followed by a markdown double bracket link tag ("[[").   Inserting spaces around the pipe allows the regex to parse the markdown text properly.

ex: | column 1 | column 2 |[[link]]|  // The link will work but it will be mistakenly rendered as part of column 2 in the HTML output.